Refactor DemoSearchInput to utilize new search entry and result structures for improved search functionality
