+++ examples/Demo/FluentUI.Demo.Client/Layout/DemoSearchInput.razor.cs
@@ -2,10 +2,7 @@
 // This file is licensed to you under the MIT License.
 // ------------------------------------------------------------------------
 
-using System.Text.RegularExpressions;
-using FluentUI.Demo.DocViewer.Models;
 using FluentUI.Demo.DocViewer.Services;
-using Markdig;
 using Microsoft.AspNetCore.Components;
 using Microsoft.FluentUI.AspNetCore.Components;
 
@@ -13,12 +10,10 @@ namespace FluentUI.Demo.Client.Layout;
 
 public partial class DemoSearchInput
 {
-    private const int MaximumDescriptionLength = 200;
-    private const int MaximumResults = 9;
-
-    private readonly List<SearchEntry> _searchEntries = [];
+    private string _currentRoute = string.Empty;
+    private readonly List<DemoSearchInputSearchEntry> _searchEntries = [];
     private string _searchText = string.Empty;
-    private SearchResult? _selectedItem;
+    private DemoSearchInputSearchResult? _selectedItem;
 
     [Inject]
     public required DocViewerService DocViewerService { get; set; }
@@ -30,137 +25,31 @@ public partial class DemoSearchInput
     {
         _searchEntries.AddRange(DocViewerService.Pages
                                                 .Where(page => !page.Hidden)
-                                                .Select(page => new SearchEntry(page)));
+                                                .Select(page => new DemoSearchInputSearchEntry(page)));
     }
 
-    private void Search(OptionsSearchEventArgs<SearchResult> args)
+    private void Search(OptionsSearchEventArgs<DemoSearchInputSearchResult> args)
     {
         _searchText = args.Text.Trim();
         if (string.IsNullOrEmpty(_searchText))
         {
-            args.Items = [];
+            args.Items = DemoSearchInputSearchResult.CreateDefaultResults(_searchEntries);
             return;
         }
 
-        args.Items = _searchEntries
-            .Select(entry => CreateSearchResult(entry, _searchText))
-            .Where(result => result is not null)
-            .OrderBy(result => result!.MatchKind)
-            .ThenBy(result => result!.MatchIndex)
-            .ThenBy(result => result!.Title, StringComparer.OrdinalIgnoreCase)
-            .Take(MaximumResults)
-            .Select(result => result!);
+        args.Items = DemoSearchInputSearchResult.CreateSearchResults(_searchEntries, _searchText);
     }
 
-    private void NavigateToSelectedPage(SearchResult? result)
+    private void NavigateToSelectedPage(DemoSearchInputSearchResult? result)
     {
         if (result is null)
         {
             return;
         }
 
+        _currentRoute = result.Route;
         _selectedItem = null;
         _searchText = string.Empty;
         NavigationManager.NavigateTo(result.Route);
-    }
-
-    private static SearchResult? CreateSearchResult(SearchEntry entry, string searchText)
-    {
-        var titleIndex = entry.Title.IndexOf(searchText, StringComparison.OrdinalIgnoreCase);
-        if (titleIndex >= 0)
-        {
-            return new SearchResult(entry.Title, entry.Route, entry.Description, null, SearchMatchKind.Title, titleIndex);
-        }
-
-        var descriptionIndex = entry.Description.IndexOf(searchText, StringComparison.OrdinalIgnoreCase);
-        if (descriptionIndex >= 0)
-        {
-            return new SearchResult(entry.Title, entry.Route, entry.Description, null, SearchMatchKind.Description, descriptionIndex);
-        }
-
-        var contentIndex = entry.Content.IndexOf(searchText, StringComparison.OrdinalIgnoreCase);
-        if (contentIndex < 0)
-        {
-            return null;
-        }
-
-        return new SearchResult(
-            entry.Title,
-            entry.Route,
-            entry.Description,
-            CreateContextSnippet(entry.Content, contentIndex, searchText.Length),
-            SearchMatchKind.Content,
-            contentIndex);
-    }
-
-    private static string CreateContextSnippet(string content, int matchIndex, int matchLength)
-    {
-        if (content.Length <= MaximumDescriptionLength)
-        {
-            return content;
-        }
-
-        var omittedPrefix = matchIndex > 0;
-        var omittedSuffix = matchIndex + matchLength < content.Length;
-        var ellipsisLength = (omittedPrefix ? 3 : 0) + (omittedSuffix ? 3 : 0);
-        var contentLength = MaximumDescriptionLength - ellipsisLength;
-        var start = Math.Max(0, matchIndex - ((contentLength - matchLength) / 2));
-        var end = Math.Min(content.Length, start + contentLength);
-
-        if (end == content.Length)
-        {
-            start = Math.Max(0, end - contentLength);
-        }
-
-        start = MoveToWordStart(content, start);
-        if (end - start > contentLength)
-        {
-            end = start + contentLength;
-        }
-
-        end = MoveToWordStart(content, end);
-        if (end <= matchIndex + matchLength)
-        {
-            end = Math.Min(content.Length, start + contentLength);
-        }
-
-        var snippet = content[start..end].Trim();
-        return $"{(start > 0 ? "..." : string.Empty)}{snippet}{(end < content.Length ? "..." : string.Empty)}";
-    }
-
-    private static string Truncate(string text, int maximumLength)
-    {
-        if (text.Length <= maximumLength)
-        {
-            return text;
-        }
-
-        var end = MoveToWordStart(text, maximumLength);
-        return $"{text[..end].TrimEnd()}...";
-    }
-
-    private static int MoveToWordStart(string text, int index)
-    {
-        while (index > 0 && index < text.Length && !char.IsWhiteSpace(text[index - 1]))
-        {
-            index--;
-        }
-
-        return index;
-    }
-
-    private static string NormalizeWhitespace(string text) => Regex.Replace(text, @"\s+", " ").Trim();
-
-    private static bool IsDocViewerDirective(string text) => text.StartsWith("{{", StringComparison.Ordinal) &&
-                                                             text.EndsWith("}}", StringComparison.Ordinal);
-
-   
-    private sealed record SearchResult(string Title, string Route, string Description, string? Context, SearchMatchKind MatchKind, int MatchIndex);
-
-    private enum SearchMatchKind
-    {
-        Title,
-        Description,
-        Content
     }
 }
